requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.8.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.2
flake8==7.1.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.10
mypy==1.11.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
thinc==8.2.5
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.5
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0